# docs/swagger.yaml
definitions:
...
  models.CreateExercisesRequest:
    properties:
      content: {}
      material_id:
        type: integer
      teacher_id:
        type: integer
      title:
        type: string
      total_marks:
        type: integer
    type: object
...
  models.Exercises:
    properties:
      content: {}
      id:
        type: integer
      material_id:
        type: integer
      teacher_id:
        type: integer
      title:
        type: string
      total_marks:
        type: integer
    type: object
...
info:
  contact: {}
  description: API documentation for Project PPL - Learnify
  title: Project PPL API
  version: 0.2.1
paths:
...
  /api/v1/exercises:
    delete:
      consumes:
      - application/json
      description: Delete an exercise by ID
      parameters:
      - description: Exercise ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Exercise
      tags:
      - Exercises
    get:
      consumes:
      - application/json
      description: Fetch all exercises for a specific material
      parameters:
      - description: Material ID
        in: query
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exercises'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exercises by Material ID
      tags:
      - Exercises
    patch:
      consumes:
      - application/json
      description: Update an existing exercise
      parameters:
      - description: Exercise ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.CreateExercisesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exercises'
      security:
      - BearerAuth: []
      summary: Update Exercise
      tags:
      - Exercises
    post:
      consumes:
      - application/json
      description: Create a new exercise
      parameters:
      - description: Exercise data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.CreateExercisesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exercises'
      security:
      - BearerAuth: []
      summary: Create Exercise
      tags:
      - Exercises
...
  /api/v1/exercises/student:
    get:
      consumes:
      - application/json
      description: Fetch all exercises for a specific material
      parameters:
      - description: Material ID
        in: query
        name: material_id
        required: true
        type: integer
      - description: Number
        in: query
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exercises'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exercises for Student
      tags:
      - Exercises
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"